+++ tests/integration/run_all.sh
@@ -113,6 +113,7 @@ ALL_CASES=(
     test_udp_sockets
     test_http_get
     test_tcp_server
+    test_tcp_x5
     test_graphics
     test_smp
     test_smp_tss